requirements.txt
# pbr should be first
pbr==5.10.0

appdirs==1.4.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
dynaconf==3.1.11
gitdb==4.0.9
GitPython==3.1.28
idna==3.4
loguru==0.6.0
pathlib==1.0.1
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.5.0
PyYAML==6.0
requests==2.28.1
rich==12.6.0
shellingham==1.5.0
smmap==5.0.0
tabulate==0.9.0
typer==0.6.1
typing_extensions==4.4.0
urllib3==1.26.12
